move URL handling functions out of utilities
