+++ tests/test_utilities.py
@@ -8,13 +8,8 @@ import pytz
 
 from requests.exceptions import HTTPError, ReadTimeout
 
-from erddapy.utilities import (
-    _clean_response,
-    check_url_response,
-    parse_dates,
-    quote_string_constraints,
-    urlopen,
-)
+from erddapy.url_handling import _clean_response, check_url_response, urlopen
+from erddapy.utilities import parse_dates, quote_string_constraints
 
 
 @pytest.mark.web
